+++ crates/astria-sequencer/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Support distributed signers as validators [#2024](https://github.com/astriaorg/astria/pull/2024)
+- Direct fetching of consensus state in `RecoverIbcClient` action [#2037](https://github.com/astriaorg/astria/pull/2037)
 
 ## [2.0.0-rc.1] - 2025-03-06